percent field added to Portfolio chart

+++ atomex/Models/Wallet.cs
@@ -9,7 +9,7 @@ namespace atomex.Models
         public decimal Amount { get; set; }
         public decimal Price { get; set; }
         public decimal Cost { get; set; }
-        public decimal PercentInPortfolio { get; set; }
+        public float PercentInPortfolio { get; set; }
         public string Address { get; set; }
         public List<Transaction> Transactions { get; set; }
     }
